pythonProject/requirements.txt
opencv-python~=4.7.0.72
numpy~=1.23.5
pandas~=2.0.1
matplotlib~=3.7.1
scikit-image~=0.20.0
imagehash~=4.3.1
pillow~=9.5.0
tensorflow~=2.12.0
scikit-learn~=1.2.2
keras~=2.12.0